feat: allow custom injectable name
The code has been refactored to allow for a custom name to be specified during the registration of an injectable. This feature is useful when you want to programmatically control the injectable naming, rather than relying on class name inference which may become unreliable after code minification. Also, fixed the test cases to check these changes correctly. Documentation is also updated to reflect these changes.